240709 [update] : upadate route handler 2
