+++ src/app/api/auth/sign-up/route.ts
@@ -2,15 +2,20 @@ import { createClient } from "@/supabase/server";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function POST(request: NextRequest) {
-  const data = await request.json();
-  const email = data.email as string;
-  const password = data.password as string;
+    const data = await request.json();
+    const email = data.email as string;
+    const password = data.password as string;
 
-  const supabase = createClient();
+    const supabase = createClient();
 
-  const {
-    data: { user },
-  } = await supabase.auth.signUp({ email, password });
+    const {
+        data: { user },
+        error,
+    } = await supabase.auth.signUp({ email, password });
 
-  return NextResponse.json(user);
+    if (error) {
+        return NextResponse.json({ error: error?.message }, { status: 401 });
+    }
+
+    return NextResponse.json(user, { status: 200 });
 }
